Created a Java EE project to test the data flow.
Specifically, the flow from app to server-side code to database and all
the way back.

So far, all it does is go from the database to a browser.

The database/schema name is "lostnfound." It has a table called "hello,"
which has one column, of type varchar. A hello message is in a row of
this table.

This commit is a squash merge from 49706cc4fba5bcc80f11e8b3b66840a5365ee9b4